Reduced code duplication to pass Sonar check
